requirements.txt
# Function dependencies, for example:
# package>=version
python-i18n[YAML]==0.3.7
pyyaml==5.3.1
requests==2.22.0
slackclient==2.4.0
oauth2client==3.0.0
firebase-admin==3.2.0